astra_xau_v2/requirements.txt
MetaTrader5
pandas
numpy
ta
requests
pymongo
python-dotenv
APScheduler
matplotlib
mplfinance
plotly
dash
rich
